fix(richtext-lexical): missing use client markers required for next.js compatibility
