+++ packages/richtext-lexical/src/field/features/Blocks/nodes/BlocksNode.tsx
@@ -1,3 +1,4 @@
+'use client'
 import type { SerializedDecoratorBlockNode } from '@lexical/react/LexicalDecoratorBlockNode'
 import type {
   DOMConversionMap,
