+++ packages/richtext-lexical/src/field/features/Link/plugins/clickableLink/index.tsx
@@ -1,0 +1,7 @@
+'use client'
+import LexicalClickableLinkPlugin from '@lexical/react/LexicalClickableLinkPlugin'
+import React from 'react'
+
+export function ClickableLinkPlugin() {
+  return <LexicalClickableLinkPlugin />
+}
