+++ packages/richtext-lexical/src/field/features/Link/plugins/floatingLinkEditor/LinkEditor/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import type { LexicalCommand } from 'lexical'
 import type { Data, Fields } from 'payload/types'
