+++ packages/richtext-lexical/src/field/features/Link/plugins/floatingLinkEditor/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import * as React from 'react'
 import { createPortal } from 'react-dom'
