+++ packages/richtext-lexical/src/field/features/Relationship/nodes/components/RelationshipComponent.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
 import { useLexicalNodeSelection } from '@lexical/react/useLexicalNodeSelection'
 import { $getNodeByKey, type ElementFormatType } from 'lexical'
