+++ packages/richtext-lexical/src/field/features/Upload/drawer/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
 import {
   $getNodeByKey,
