+++ packages/richtext-lexical/src/field/features/lists/CheckList/index.ts
@@ -1,18 +1,18 @@
 import { INSERT_CHECK_LIST_COMMAND, ListItemNode, ListNode } from '@lexical/list'
-import { CheckListPlugin } from '@lexical/react/LexicalCheckListPlugin'
 
 import type { FeatureProvider } from '../../types'
 
-import { SlashMenuOption } from '../../../lexical/plugins/SlashMenu/LexicalTypeaheadMenuPlugin/LexicalMenu'
+import { SlashMenuOption } from '../../../lexical/plugins/SlashMenu/LexicalTypeaheadMenuPlugin/types'
 import { ChecklistIcon } from '../../../lexical/ui/icons/Checklist'
 import { ListHTMLConverter, ListItemHTMLConverter } from '../htmlConverter'
 import { CHECK_LIST } from './markdownTransformers'
+import { LexicalCheckListPlugin } from './plugin'
 
 // 345
 // carbs 7
 export const CheckListFeature = (): FeatureProvider => {
   return {
-    feature: ({ featureProviderMap, resolvedFeatures, unsanitizedEditorConfig }) => {
+    feature: ({ featureProviderMap }) => {
       return {
         markdownTransformers: [CHECK_LIST],
         nodes:
@@ -36,7 +36,7 @@ export const CheckListFeature = (): FeatureProvider => {
               ],
         plugins: [
           {
-            Component: CheckListPlugin,
+            Component: LexicalCheckListPlugin,
             position: 'normal',
           },
         ],
