+++ packages/richtext-lexical/src/field/features/lists/CheckList/plugin/index.tsx
@@ -1,0 +1,7 @@
+'use client'
+import { CheckListPlugin } from '@lexical/react/LexicalCheckListPlugin'
+import React from 'react'
+
+export function LexicalCheckListPlugin() {
+  return <CheckListPlugin />
+}
