+++ packages/richtext-lexical/src/field/features/lists/UnorderedList/index.ts
@@ -1,11 +1,11 @@
 import { INSERT_UNORDERED_LIST_COMMAND, ListItemNode, ListNode } from '@lexical/list'
-import { ListPlugin } from '@lexical/react/LexicalListPlugin'
 
 import type { FeatureProvider } from '../../types'
 
-import { SlashMenuOption } from '../../../lexical/plugins/SlashMenu/LexicalTypeaheadMenuPlugin/LexicalMenu'
+import { SlashMenuOption } from '../../../lexical/plugins/SlashMenu/LexicalTypeaheadMenuPlugin/types'
 import { UnorderedListIcon } from '../../../lexical/ui/icons/UnorderedList'
 import { ListHTMLConverter, ListItemHTMLConverter } from '../htmlConverter'
+import { LexicalListPlugin } from '../plugin'
 import { UNORDERED_LIST } from './markdownTransformer'
 
 export const UnoderedListFeature = (): FeatureProvider => {
@@ -31,7 +31,7 @@ export const UnoderedListFeature = (): FeatureProvider => {
         ],
         plugins: [
           {
-            Component: ListPlugin,
+            Component: LexicalListPlugin,
             position: 'normal',
           },
         ],
