+++ packages/richtext-lexical/src/field/features/lists/plugin/index.tsx
@@ -1,0 +1,7 @@
+'use client'
+import { ListPlugin } from '@lexical/react/LexicalListPlugin'
+import React from 'react'
+
+export function LexicalListPlugin() {
+  return <ListPlugin />
+}
