+++ packages/richtext-lexical/src/field/lexical/config/EditorConfigProvider.tsx
@@ -1,3 +1,4 @@
+'use client'
 import * as React from 'react'
 import { createContext, useContext, useEffect, useMemo, useState } from 'react'
